% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dissolved_organic_carbon}
\alias{dissolved_organic_carbon}
\title{Dissolved organic carbon}
\format{A data.frame with 4 variables: \code{year}, \code{watershed},
\code{concentration} (Mg/L), and \code{load} (kg/ha)}
\usage{
dissolved_organic_carbon
}
\description{
Concentration and load for dissolved organic carbon by year on each STRIPS
watershed
}
\examples{
dissolved_organic_carbon
}
\keyword{datasets}

